refactor: flatten shared/ single-file subdirs and tidy components/
Several shared/ subdirs held a single file as aspirational namespaces
that never grew. Pull them into existing homes where the file actually
fits, and pair the marquee hook + overlay (a single logical unit that
was split across two trees).

shared/ flattening:
- shared/animation/easing.ts        → shared/utils/easing.ts
- shared/async/async-utils.ts       → shared/utils/async-utils.ts
- shared/events/domain-events.ts    → shared/utils/domain-events.ts
- shared/media/ac3-decoder.ts       → shared/utils/ac3-decoder.ts
- shared/hooks/use-marquee-selection.ts + components/marquee-overlay.tsx
                                    → shared/marquee/*  (paired unit)
- shared/components/color-scopes-view.tsx
                                    → features/preview/components/  (its only consumer)

components/ tidying — split mixed app-shell vs shared UI:
- components/error-boundary.tsx     → app/error-boundary.tsx
- components/pwa-install-prompt.tsx → app/pwa-install-prompt.tsx

src/components/ is now just shadcn ui/ and brand/.

.oxlintrc.json: drop the now-stale shared/animation/** path from the
framework-agnostic rule (timeline/ and projects/ remain).

All 2268 tests pass; lint, feature-boundary, legacy-lib, tsc all clean.
