+++ src/app.tsx
@@ -3,8 +3,8 @@ import { RouterProvider, createRouter } from '@tanstack/react-router'
 import { TooltipProvider } from '@/components/ui/tooltip'
 import { GlobalTooltip } from '@/components/ui/global-tooltip'
 import { Toaster } from '@/components/ui/sonner'
-import { ErrorBoundary } from '@/components/error-boundary'
-import { PwaInstallPrompt } from '@/components/pwa-install-prompt'
+import { ErrorBoundary } from '@/app/error-boundary'
+import { PwaInstallPrompt } from '@/app/pwa-install-prompt'
 import { WorkspaceGate } from '@/features/workspace-gate'
 import { routeTree } from './routeTree.gen'
