+++ src/features/editor/components/editor.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { createLogger } from '@/shared/logging/logger'
 import { i18n } from '@/i18n'
 import { ResizablePanelGroup, ResizablePanel, ResizableHandle } from '@/components/ui/resizable'
-import { ErrorBoundary } from '@/components/error-boundary'
+import { ErrorBoundary } from '@/app/error-boundary'
 import { Toolbar } from './toolbar'
 import { MediaSidebar } from './media-sidebar'
 import { PropertiesSidebar } from './properties-sidebar'
