+++ src/features/editor/components/preview-area.tsx
@@ -19,7 +19,7 @@ import { useEditorStore } from '@/shared/state/editor'
 import { EDITOR_LAYOUT_CSS_VALUES, getEditorLayout } from '@/config/editor-layout'
 import { InteractionLockRegion } from './interaction-lock-region'
 import { Button } from '@/components/ui/button'
-import { ErrorBoundary } from '@/components/error-boundary'
+import { ErrorBoundary } from '@/app/error-boundary'
 
 interface PreviewAreaProps {
   project: {
