+++ src/features/media-library/components/media-library.tsx
@@ -62,7 +62,7 @@ import {
 } from '@/components/ui/dialog'
 import { Collapsible, CollapsibleContent, CollapsibleTrigger } from '@/components/ui/collapsible'
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip'
-import { MarqueeOverlay } from '@/components/marquee-overlay'
+import { MarqueeOverlay } from '@/shared/marquee/marquee-overlay'
 import { cn } from '@/shared/ui/cn'
 import { MediaGrid } from './media-grid'
 import { CompositionsSection } from './compositions-section'
@@ -97,7 +97,7 @@ import {
   isMarqueeJustFinished,
   useMarqueeSelection,
   type MarqueeItem,
-} from '@/shared/hooks/use-marquee-selection'
+} from '@/shared/marquee/use-marquee-selection'
 
 function CopyButton({ text }: { text: string }) {
   const { t } = useTranslation()
