+++ src/features/preview/components/color-scopes-view.tsx
