+++ src/features/timeline/components/timeline-media-drop-zone.tsx
@@ -26,7 +26,7 @@ import {
   findNearestAvailableSpaceInTrackItems,
   type CollisionRect,
 } from '../utils/collision-utils'
-import { mapWithConcurrency } from '@/shared/async/async-utils'
+import { mapWithConcurrency } from '@/shared/utils/async-utils'
 import { useExternalDragPreview } from '../hooks/use-external-drag-preview'
 import { useCompositionNavigationStore } from '../stores/composition-navigation-store'
 import { wouldCreateCompositionCycle } from '../utils/composition-graph'
