+++ src/features/timeline/stores/actions/shared.ts
@@ -6,7 +6,7 @@ import type { TimelineItem, TimelineTrack } from '@/types/timeline'
 import type { Transition } from '@/types/transition'
 import type { ItemKeyframes } from '@/types/keyframe'
 import { createLogger } from '@/shared/logging/logger'
-import { emitDomainEvent } from '@/shared/events/domain-events'
+import { emitDomainEvent } from '@/shared/utils/domain-events'
 import { useTimelineCommandStore } from '../timeline-command-store'
 import { useItemsStore } from '../items-store'
 import { useTransitionsStore } from '../transitions-store'
