+++ src/runtime/composition-runtime/utils/audio-decode-cache.test.ts
@@ -143,7 +143,7 @@ vi.mock('@/infrastructure/storage', () => ({
   ...decodedPreviewAudioMocks,
   ...mediaDbMocks,
 }))
-vi.mock('@/shared/media/ac3-decoder', () => ac3Mocks)
+vi.mock('@/shared/utils/ac3-decoder', () => ac3Mocks)
 vi.mock('@/infrastructure/browser/object-url-registry', () => objectUrlRegistryMocks)
 vi.mock('./preview-audio-conform', () => previewAudioConformMocks)
 vi.mock('mediabunny', () => mediabunnyMocks)
