+++ src/runtime/composition-runtime/utils/audio-decode-cache.ts
@@ -23,7 +23,7 @@ import {
   deleteDecodedPreviewAudio,
   getMedia,
 } from '@/infrastructure/storage'
-import { ensureAc3DecoderRegistered, isAc3AudioCodec } from '@/shared/media/ac3-decoder'
+import { ensureAc3DecoderRegistered, isAc3AudioCodec } from '@/shared/utils/ac3-decoder'
 import type { DecodedPreviewAudioMeta, DecodedPreviewAudioBin } from '@/types/storage'
 import { persistPreviewAudioConform } from './preview-audio-conform'
