+++ src/shared/marquee/use-marquee-selection.ts
