+++ src/shared/utils/easing.ts
